+++ src/index.html
@@ -4,7 +4,7 @@
   </head>
   <body>
 
-    <div id='root'></div>
+    <div id='react-root'></div>
 
     <div id="webpack-samples">
     </div>
